split pixel_simd.h into several files
