+++ dotproduct/pixel_simd_sse.h
@@ -1,0 +1,6 @@
+#ifndef PIXEL_SIMD_SSE_H
+#define PIXEL_SIMD_SSE_H
+
+#include "pixel_simd_base.h"
+
+#endif
